Adding horizontal scrolling
Adding Horizontal Scrolling via shift-scroll or by swapping the scroll orientation in the menu per #30

Also breaks the SWIFT code out to modules for easier management.
