+++ Sources/PowerMateAgent/LEDControl.swift
@@ -1,0 +1,46 @@
+import Foundation
+import AppKit
+
+// MARK: - LED queue
+
+// LED updates run on a background queue so they never block the main run loop (where HID
+// reports are delivered); blocking the main loop was causing scroll to pause then resume.
+let ledQueue = DispatchQueue(label: "com.powermate.agent.led", qos: .utility)
+
+func setLEDOffMain(_ value: UInt8) {
+    ledQueue.async { _ = driver.setLEDBrightness(value) }
+}
+
+// MARK: - Throb animation
+
+// Optional LED feedback: dim when idle, throb while turning, full brightness when button held.
+var throbTimer: DispatchSourceTimer?
+var smoothedThrobBrightness: Double = 80
+private let throbPeriod: CFTimeInterval = 1.5
+private let throbIdleTimeout: CFTimeInterval = 0.5
+private let throbTick: CFTimeInterval = 0.025
+private let throbSmooth: Double = 0.22
+
+func startThrob() {
+    lastRotationTime = CFAbsoluteTimeGetCurrent()
+    guard throbTimer == nil, !audioMeterActive else { return }
+    throbTimer = DispatchSource.makeTimerSource(queue: .main)
+    throbTimer?.schedule(deadline: .now(), repeating: throbTick)
+    throbTimer?.setEventHandler {
+        guard driver.isConnected, !isButtonDown else { return }
+        let now = CFAbsoluteTimeGetCurrent()
+        if now - lastRotationTime > throbIdleTimeout {
+            throbTimer?.cancel()
+            throbTimer = nil
+            setLEDOffMain(80)
+            smoothedThrobBrightness = 80
+            return
+        }
+        let t = now.truncatingRemainder(dividingBy: throbPeriod) / throbPeriod
+        let target = 80 + 175 * (0.5 + 0.5 * sin(t * 2 * .pi))
+        smoothedThrobBrightness += (target - smoothedThrobBrightness) * throbSmooth
+        let value = UInt8(max(0, min(255, smoothedThrobBrightness.rounded())))
+        setLEDOffMain(value)
+    }
+    throbTimer?.resume()
+}
